fix: block trace hang in UT, we move it to example
